Fix Star Harvester syncing issues

+++ src/main/java/tterrag/supermassivetech/util/Utils.java
@@ -257,7 +257,7 @@ public class Utils
         if (FMLCommonHandler.instance().getEffectiveSide() == Side.SERVER && showParticles && FMLClientHandler.instance().getClient().effectRenderer != null
                 && Minecraft.getMinecraft().thePlayer != null)
             FMLClientHandler.instance().getClient().effectRenderer.addEffect(new EntityCustomSmokeFX(Minecraft.getMinecraft().thePlayer.worldObj, entity.posX, entity.posY, entity.posZ, xCoord + 0.5,
-                    yCoord + 0.5, zCoord + 0.5, (double) (1 / range)));
+                    yCoord + 0.5, zCoord + 0.5, 1 / range));
     }
 
     /**
